fixed current directory problem. added healthy does of paranoia to file_rel_chdir, and removed bogus chdir from main.cpp (from test code)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@238 3db68df2-c116-0410-a063-a993310a9797

+++ source/terrain/TextureManager.cpp
@@ -23,7 +23,7 @@ void CTextureManager::AddTextureType(const char* name)
 	m_TerrainTextures.resize(m_TerrainTextures.size()+1);
 	STextureType& ttype=m_TerrainTextures.back();
 	ttype.m_Name=name;
-	ttype.m_Index=m_TerrainTextures.size()-1;
+	ttype.m_Index=(int)m_TerrainTextures.size()-1;
 }
 
 CTextureEntry* CTextureManager::FindTexture(const char* filename)
